Sets the docusaurus trailingSlash property to true and adapt site-internal links (#21)
Why this change? The webserver on eclipse.dev will 301-forward URLs without a trailing slash to URLs with a trailing slash, probably for SEO reasons. Example: https://eclipse.dev/lmos/arc is forwarded to https://eclipse.dev/lmos/arc/. This causes problems with some of the site-internal links created by docusaurus, leading to 404, as described in https://github.com/eclipse-lmos/website/issues/20. Setting the trailingSlash property to true will make docusaurus add trailing slashes appropriately to links. URLs without a trailing slash are forwarded to the start page of the documentation (but users should not get to such URLs anyways, because the webserver forwards to a URL with trailing slash).

Fixes #20